Update resolver internals docs
Since the resolver internals docs were written, we added a lot more features to the resolver, which should be documented.

As usual, these docs are not targeted at regular users, but should give interested readers an insight into the internals of uv and help advanced users with especially hard resolver problems.