Errors for builds in non-build and workspace root pyproject.toml
When running `uv pip install .` in a directory with a pyproject.toml that does not configure a build, we currently invoke setuptools and get a wheel we can't parse (https://github.com/astral-sh/uv/issues/11344).

Instead, we reject such builds altogether and require that there is at least a metadata source if no build backend has been defined. As a special case, we hint when the build failed on a workspace root, which easily happens when running `uv pip install .` in a workspace root, assuming it would be a `uv sync --all-packages`.