Run uv-dev tests
In #6827, we switched the uv-dev binary to not being built by default. As an unintended side effect, we were also stopping to run the tests that ensured the schema was up-to-date.

To fix this, we split uv-dev into an unconditional library, with only the binary being a conditional build. This way, `cargo test` and `cargo nextest` pick those tests up again.

An alternative would be running tests with the `dev` feature, with the side effect of always building the uv-dev binary, too.