Reduce edge hashing
Building the `HashSet` was a block in the profile (2.6% of the resolver thread). We don't need to hash at all since we're deduplicating edges when reading them anyway.

The benchmark isn't too clear but the profile show it as an improvement.

```
Benchmark 1: ./uv-main pip compile --no-progress scripts/requirements/airflow.in --universal
  Time (mean ± σ):     102.0 ms ±   6.8 ms    [User: 116.4 ms, System: 114.6 ms]
  Range (min … max):    93.1 ms … 115.8 ms    32 runs

Benchmark 2: ./uv-branch pip compile --no-progress scripts/requirements/airflow.in --universal
  Time (mean ± σ):      98.6 ms ±   4.4 ms    [User: 114.0 ms, System: 112.7 ms]
  Range (min … max):    93.3 ms … 112.4 ms    31 runs

Summary
  ./uv-branch pip compile --no-progress scripts/requirements/airflow.in --universal ran
    1.03 ± 0.08 times faster than ./uv-main pip compile --no-progress scripts/requirements/airflow.in --universal
```